pyproject.toml
[tool.poetry]
name = "autofillpdffiles"
version = "0.1.0"
description = ""
authors = ["vaduska <vadimsalapugin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.8.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
